feat : 채팅 내역 응답값에 상대방 정보 추가 - 프론트 요청으로 채팅 내역 응답값에 상대방 정보 추가
