+++ chat-server/src/main/java/com/kuddy/chatserver/chat/controller/ChatRoomController.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 import javax.validation.Valid;
 
+import com.kuddy.chatserver.chat.service.ChatRoomService;
 import org.springframework.http.ResponseEntity;
 
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -17,8 +18,8 @@ import com.kuddy.chatserver.chat.dto.request.ChatReqDto;
 import com.kuddy.chatserver.chat.dto.response.ChatHistoryResDto;
 import com.kuddy.chatserver.chat.dto.response.ChatRoomListResDto;
 import com.kuddy.chatserver.chat.dto.response.ChatRoomResDto;
-import com.kuddy.chatserver.chat.service.ChatRoomService;
-import com.kuddy.chatserver.chat.service.ChatService;
+import com.kuddy.chatserver.chat.service.ChatRoomConnectInfoService;
+import com.kuddy.chatserver.chat.service.ChattingService;
 import com.kuddy.common.member.domain.Member;
 import com.kuddy.common.response.StatusEnum;
 import com.kuddy.common.response.StatusResponse;
@@ -30,7 +31,8 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/chat/v1/chatrooms")
 @RequiredArgsConstructor
 public class ChatRoomController {
-	private final ChatService chatService;
+	private final ChattingService chattingService;
+	private final ChatRoomConnectInfoService chatRoomConnectInfoService;
 	private final ChatRoomService chatRoomService;
 	private static final String CHAT_ROOM_DISCONNECTED= "성공적으로 접속 해제했습니다.";
 
@@ -40,8 +42,8 @@ public class ChatRoomController {
 	public ResponseEntity<StatusResponse> createChatRoom(@RequestBody @Valid final ChatReqDto requestDto, @AuthUser Member member) {
 
 		// 채팅방을 만들어준다.
-		Long roomId = chatService.makeChatRoom(member, requestDto);
-		Room room = chatService.findByRoomId(roomId);
+		Long roomId = chatRoomService.makeChatRoom(member, requestDto);
+		Room room = chatRoomService.findByRoomId(roomId);
 		ChatRoomResDto response = new ChatRoomResDto(room);
 		URI location = ServletUriComponentsBuilder.fromCurrentRequest()
 			.path("/{id}")
@@ -60,7 +62,7 @@ public class ChatRoomController {
 	@GetMapping
 	@PreAuthorize("isAuthenticated()")
 	public ResponseEntity<StatusResponse> chatRoomList(@AuthUser Member member) {
-		List<ChatRoomListResDto> response = chatService.getChatList(member);
+		List<ChatRoomListResDto> response = chatRoomService.getChatList(member);
 		return ResponseEntity.ok(StatusResponse.builder()
 				.status(StatusEnum.OK.getStatusCode())
 				.message(StatusEnum.OK.getCode())
@@ -72,9 +74,9 @@ public class ChatRoomController {
 	@GetMapping("/{roomId}")
 	@PreAuthorize("isAuthenticated()")
 	public ResponseEntity<StatusResponse> chattingList(@PathVariable("roomId") Long roomId, @AuthUser Member member) {
-		String email = chatService.checkRoomIdOwnerValidation(member, roomId);
-		chatService.updateCountAllZero(roomId, email);
-		ChatHistoryResDto response = chatService.getChattingList(roomId, member);
+		String email = chatRoomService.checkRoomIdOwnerValidation(member, roomId);
+		chatRoomService.updateCountAllZero(roomId, email);
+		ChatHistoryResDto response = chattingService.getChattingList(roomId, member);
 		return ResponseEntity.ok(StatusResponse.builder()
 			.status(StatusEnum.OK.getStatusCode())
 			.message(StatusEnum.OK.getCode())
@@ -86,7 +88,7 @@ public class ChatRoomController {
 	@GetMapping("/check")
 	@PreAuthorize("isAuthenticated()")
 	public ResponseEntity<StatusResponse> hasChatRoom(@AuthUser Member member,@RequestParam("email") String email) {
-		Room room = chatService.findByMembers(member, email);
+		Room room = chatRoomService.findByMembers(member, email);
 		ChatRoomResDto response = new ChatRoomResDto(room);
 		return ResponseEntity.ok(StatusResponse.builder()
 				.status(StatusEnum.OK.getStatusCode())
@@ -101,9 +103,9 @@ public class ChatRoomController {
 	@PreAuthorize("isAuthenticated()")
 	public ResponseEntity<StatusResponse> disconnectChat(@PathVariable("roomId") Long roomId,
 		@RequestParam("email") String email, @AuthUser Member member) {
-		String loginEmail = chatService.checkRoomIdOwnerValidation(member, roomId);
-		chatService.checkEmailValidation(loginEmail, email);
-		chatRoomService.disconnectChatRoom(roomId, loginEmail);
+		String loginEmail = chatRoomService.checkRoomIdOwnerValidation(member, roomId);
+		chatRoomService.checkEmailValidation(loginEmail, email);
+		chatRoomConnectInfoService.disconnectChatRoom(roomId, loginEmail);
 		return ResponseEntity.ok(StatusResponse.builder()
 			.status(StatusEnum.OK.getStatusCode())
 			.message(StatusEnum.OK.getCode())
@@ -115,7 +117,7 @@ public class ChatRoomController {
 	// 메시지 전송 후 callback
 	@PostMapping("/chatrooms/callback")
 	public ResponseEntity<StatusResponse> sendNotification(@Valid @RequestBody Message message) {
-		Message savedMessage = chatService.sendNotificationAndSaveMessage(message);
+		Message savedMessage = chattingService.sendNotificationAndSaveMessage(message);
 		return ResponseEntity.ok(StatusResponse.builder()
 			.status(StatusEnum.OK.getStatusCode())
 			.message(StatusEnum.OK.getCode())
