+++ chat-server/src/main/java/com/kuddy/chatserver/chat/controller/ChattingController.java
@@ -1,6 +1,6 @@
 package com.kuddy.chatserver.chat.controller;
 
-import com.kuddy.chatserver.chat.service.ChatService;
+import com.kuddy.chatserver.chat.service.ChattingService;
 import com.kuddy.common.chat.domain.Message;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.handler.annotation.Header;
@@ -15,16 +15,16 @@ import java.io.IOException;
 @RequestMapping("/chat/v1")
 @RequiredArgsConstructor
 public class ChattingController {
-    private final ChatService chatService;
+    private final ChattingService chattingService;
 
     @MessageMapping("/message")
     public void sendMessage(@Valid Message message, @Header("Authorization") final String accessToken) {
-        chatService.sendMessage(message, accessToken);
+        chattingService.sendMessage(message, accessToken);
     }
 
     @MessageMapping("/updateMessage")
     public void updateMessage(@Valid Message updatedMessage, @Header("Authorization") final String accessToken) throws IOException {
         // 데이터베이스 상태 업데이트 또는 서비스 로직 실행
-        chatService.updateMessage(updatedMessage, accessToken);
+        chattingService.updateMessage(updatedMessage, accessToken);
     }
 }
